place Bunch isolated to avoid importing full misc
